feat: introduce React Router with landing page
