+++ package.json
@@ -30,8 +30,10 @@
     "pixi.js": "^7.3.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-helmet-async": "^2.0.5",
     "react-hook-form": "^7.49.2",
     "react-icons": "^5.5.0",
+    "react-router-dom": "^7.6.3",
     "tailwind-merge": "^2.6.0",
     "tailwindcss-animate": "^1.0.7",
     "tonal": "^6.4.2",
